+++ src/allocator.cpp
@@ -170,7 +170,7 @@ namespace libtorrent
 		// make the two surrounding pages non-readable and -writable
 		mprotect(block - page, page, PROT_READ | PROT_WRITE);
 		alloc_header* h = reinterpret_cast<alloc_header*>(block - page);
-		const int num_pages = (h->size + (page-1)) / page + 2;
+		const int num_pages = int((h->size + (page-1)) / page + 2);
 		TORRENT_ASSERT(h->magic == 0x1337);
 		mprotect(block + (num_pages-2) * page, page, PROT_READ | PROT_WRITE);
 //		std::fprintf(stderr, "free: %p head: %p tail: %p size: %d\n", block, block - page, block + h->size, int(h->size));
